+++ sqlalchemy_cubrid/__init__.py
@@ -49,7 +49,7 @@ from sqlalchemy.sql.sqltypes import (
     TIMESTAMP,
 )
 
-__version__ = "2.0.0"
+__version__ = "2.1.0"
 
 __all__ = (
     "insert",