+++ test/conftest.py
@@ -14,6 +14,7 @@ if "--dburi" in sys.argv or any(a.startswith("--dburi=") for a in sys.argv):
 
     registry.register("cubrid", "sqlalchemy_cubrid.dialect", "CubridDialect")
     registry.register("cubrid.cubrid", "sqlalchemy_cubrid.dialect", "CubridDialect")
+    registry.register("cubrid.pycubrid", "sqlalchemy_cubrid.pycubrid_dialect", "PyCubridDialect")
 
     import pytest